ListMap (List<Map>, duh) and first implementation FixedListMap
Initial thoughts of a new data structure that maintains a list of
mappings. Obviously one can mimic it using dozens of individual List and
Map implementations. Here there is an idea of a low memory footprint
implementation: a table, whose lines as indexed by number and columns by
a key (of any type).

FixedListMap is fixed in both senses: the number of lines cannot grow
nor shrink. The mappings themselves are also fixed: one cannot put a new
key neither remove an existing one. It is not fully tested yet.

Planned implementations:
- one that extends FixedListMap allowing removal of keys (but not adding
an unseen key);
- one that extends the previous one allowing full semantics of
List<Map>, that is, adding and removing items of outer List AND adding
new keys and removal of keys. Optionally will have a compact() method,
which will compact in both dimensions.