ByteTrie: a trie of byte sequences (first thoughts)
ArrayByteTrie is the most naive implementation thereof. So naive, so
simple that can be used as a known good implementation (of course,
this simplicity does not preclude testing).

ByteTries are not intended to be used per se, but as a helper to
implement very big and very fast Set<String> and Map<String, V>.