FixedKeysListMap, List<Map> whose size can freely change
The key set is fixed at creation time, though.

Note that all returned values are online. That means, for example, that
remove() will return a Map<K, V> whose contents, when queried, will be
that of next element (or will throw exception if the last element was
removed); this example shows one of, perhaps, many violations of List
interface.

Users should be advised specifically to no use any returned Map after
any modification. Maybe later this will be explicitly coded (throwing a
ConcurrentModificationException).

[for the given example, note that the returned Map is produced before
removal; so, from the beginning that Map is doomed]