committing to use java 5
There was actually a dependency on java 1.6: some tests were using
AbstractMap.SimpleEntry, which is public on java 6, but still package
private on java 5.

Since it was a single dependency and was located among unit tests, it
was preferrable to copy the code (with proper copyright notice) than
bump JDK version. When a true dependency on java 6 (by true I mean a non
trivial or among main code) arises, this copy should be extirpated in
favor of the oficial one.