structure tests of MultiDimensionalArrayAbstractUnitTest
MultiDimensionalArrayAbstractUnitTest proper tests checks invariants for
any implementation of MultiDimensionalArray. Subsuites of it will care
for special invariants that produces other MultiDimensionalArray
(clones, copy constructors and alike). For the time being these are
those subclasses:
* $ConstructorFromItself (copy constructor);
* $ConstructorFromKnownGoodImplementation (copy constructor);
* $ConstructorFromRawModel (initialization constructor).

This structure is not as good as it should be. Their clients (i.e. the
concrete tests that inherits therefrom) still have the duty of extending
the subsuites. An optimal solution would have both qualities:
* each invariant should be a test *per se*, so Eclipse (and maven
  surefire) could list EXACTLY what is being tested;
* each concrete test should only extend the main suite and have the
  aforementioned structure for free.